split up in multiple files and restructured code
