+++ src/_template.html
@@ -18,54 +18,5 @@
         </button>
       </div>
     </aside>
-    <!--aside class="prov_right" >
-        <div class="shortcuts">
-            <div>
-                <button id="undoStep" title="undo" class="btn btn-default btn-sm"><i class="fa fa-undo"></i></button>
-                <button data-toggle="modal" data-target="#attachNote" title="add note" class="btn btn-default btn-sm"><i class="fa fa-sticky-note-o"></i></button>
-                <button id="bookmarkState" title="bookmark state" class="btn btn-default btn-sm"><i class="fa fa-bookmark"></i></button>
-                <button id="attachScreenshot" title="add screenshot" class="btn btn-default btn-sm"><i class="fa fa-camera"></i></button>
-            </div>
-        </div>
-    </aside>-->
-  </div>
-</div>
-
-<!--login dialog-->
-<div class="modal fade" id="loginDialog" tabindex="-1" role="dialog" aria-labelledby="loginDialog" data-keyboard="false" data-backdrop="static">
-  <div class="modal-dialog modal-sm">
-    <div class="modal-content">
-      <div class="modal-header">
-        <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span
-          aria-hidden="true">&times;</span></button>
-        <h4 class="modal-title">Please login</h4>
-      </div>
-      <div class="modal-body">
-
-      </div>
-    </div>
-  </div>
-</div>
-
-<!--attach note dialog-->
-<div class="modal fade" id="attachNote" tabindex="-1" role="dialog">
-  <div class="modal-dialog modal-sm">
-    <div class="modal-content">
-      <div class="modal-header">
-        <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span
-          aria-hidden="true">&times;</span></button>
-        <h4 class="modal-title">Attach Note</h4>
-      </div>
-      <div class="modal-body">
-        <form>
-          <div class="form-group">
-            <label for="attachNote_note">Note</label>
-            <textarea class="form-control" id="attachNote_note" rows="5" placeholder="Description"
-                      required="required"></textarea>
-          </div>
-          <button type="submit" class="btn btn-default"><i class="fa fa-save"></i> Save</button>
-        </form>
-      </div>
-    </div>
   </div>
 </div>
